third_party/blink/web_tests/external/wpt/html/semantics/forms/constraints/form-validation-validity-valueMissing-expected.txt

This is a testharness.js-based test.
Found 96 tests; 94 PASS, 2 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS [INPUT in TEXT status] The required attribute is not set
PASS [INPUT in TEXT status] The value is not empty and required is true
PASS [INPUT in TEXT status] The value is empty and required is true
PASS [INPUT in SEARCH status] The required attribute is not set
PASS [INPUT in SEARCH status] The value is not empty and required is true
PASS [INPUT in SEARCH status] The value is empty and required is true
PASS [INPUT in TEL status] The required attribute is not set
PASS [INPUT in TEL status] The value is not empty and required is true
PASS [INPUT in TEL status] The value is empty and required is true
PASS [INPUT in URL status] The required attribute is not set
PASS [INPUT in URL status] The value is not empty and required is true
PASS [INPUT in URL status] The value is empty and required is true
PASS [INPUT in EMAIL status] The required attribute is not set
PASS [INPUT in EMAIL status] The value is not empty and required is true
PASS [INPUT in EMAIL status] The value is empty and required is true
PASS [INPUT in PASSWORD status] The required attribute is not set
PASS [INPUT in PASSWORD status] The value is not empty and required is true
PASS [INPUT in PASSWORD status] The value is empty and required is true
PASS [INPUT in DATETIME-LOCAL status] The required attribute is not set
PASS [INPUT in DATETIME-LOCAL status] Valid local date and time string(2000-12-10T12:00:00)
FAIL [INPUT in DATETIME-LOCAL status] Valid local date and time string(2000-12-10 12:00) assert_false: The validity.valueMissing should be false. expected false got true
PASS [INPUT in DATETIME-LOCAL status] Valid local date and time string(1979-10-14T12:00:00.001)
PASS [INPUT in DATETIME-LOCAL status] The value attribute is a number(1234567)
PASS [INPUT in DATETIME-LOCAL status] The value attribute is a Date object
PASS [INPUT in DATETIME-LOCAL status] Invalid local date and time string(1979-10-99 99:99)
FAIL [INPUT in DATETIME-LOCAL status] Valid local date and time string(1979-10-14 12:00:00) assert_false: The validity.valueMissing should be false. expected false got true
PASS [INPUT in DATETIME-LOCAL status] Invalid local date and time string(2001-12-21  12:00)-two white space
PASS [INPUT in DATETIME-LOCAL status] the value attribute is a string(abc)
PASS [INPUT in DATETIME-LOCAL status] The value attribute is empty string
PASS [INPUT in DATE status] The required attribute is not set
PASS [INPUT in DATE status] Valid date string(2000-12-10)
PASS [INPUT in DATE status] Valid date string(9999-01-01)
PASS [INPUT in DATE status] The value attribute is a number(1234567)
PASS [INPUT in DATE status] The value attribute is a Date object
PASS [INPUT in DATE status] Invalid date string(9999-99-99)
PASS [INPUT in DATE status] Invalid date string(37-01-01)
PASS [INPUT in DATE status] Invalid date string(2000/01/01)
PASS [INPUT in DATE status] The value attribute is empty string
PASS [INPUT in MONTH status] The required attribute is not set
PASS [INPUT in MONTH status] Valid month string(2000-12)
PASS [INPUT in MONTH status] Valid month string(9999-01)
PASS [INPUT in MONTH status] The value attribute is a number(1234567)
PASS [INPUT in MONTH status] The value attribute is a Date object
PASS [INPUT in MONTH status] Invalid month string(2000-99)
PASS [INPUT in MONTH status] Invalid month string(37-01)
PASS [INPUT in MONTH status] Invalid month string(2000/01)
PASS [INPUT in MONTH status] The value attribute is empty string
PASS [INPUT in WEEK status] The required attribute is not set
PASS [INPUT in WEEK status] Valid week string(2000-W12)
PASS [INPUT in WEEK status] Valid week string(9999-W01)
PASS [INPUT in WEEK status] The value attribute is a number(1234567)
PASS [INPUT in WEEK status] The value attribute is a Date object
PASS [INPUT in WEEK status] Invalid week string(2000-W99)
PASS [INPUT in WEEK status] invalid week string(2000-W00)
PASS [INPUT in WEEK status] invalid week string(2000-w01)
PASS [INPUT in WEEK status] The value attribute is empty string
PASS [INPUT in TIME status] The required attribute is not set
PASS [INPUT in TIME status] Validtime string(12:00:00)
PASS [INPUT in TIME status] Validtime string(12:00)
PASS [INPUT in TIME status] Valid time string(12:00:60.001)
PASS [INPUT in TIME status] Valid time string(12:00:60.01)
PASS [INPUT in TIME status] Valid time string(12:00:60.1)
PASS [INPUT in TIME status] The value attribute is a number(1234567)
PASS [INPUT in TIME status] The value attribute is a time object
PASS [INPUT in TIME status] Invalid time string(25:00:00)
PASS [INPUT in TIME status] Invalid time string(12:60:00)
PASS [INPUT in TIME status] Invalid time string(12:00:60)
PASS [INPUT in TIME status] Invalid time string(12:00:00:001)
PASS [INPUT in TIME status] The value attribute is empty string
PASS [INPUT in NUMBER status] The required attribute is not set
PASS [INPUT in NUMBER status] Value is an integer with a leading symbol '+'
PASS [INPUT in NUMBER status] Value is a number with a '-' symbol
PASS [INPUT in NUMBER status] Value is a number in scientific notation form(e is in lowercase)
PASS [INPUT in NUMBER status] Value is a number in scientific notation form(E is in uppercase)
PASS [INPUT in NUMBER status] Value is -0
PASS [INPUT in NUMBER status] Value is a number with some white spaces
PASS [INPUT in NUMBER status] Value is Math.pow(2, 1024)
PASS [INPUT in NUMBER status] Value is Math.pow(-2, 1024)
PASS [INPUT in NUMBER status] Value is a string that cannot be converted to a number
PASS [INPUT in NUMBER status] The value attribute is empty string
PASS [INPUT in CHECKBOX status] The required attribute is not set
PASS [INPUT in CHECKBOX status] The checked attribute is true
PASS [INPUT in CHECKBOX status] The checked attribute is false
PASS [INPUT in RADIO status] The required attribute is not set
PASS [INPUT in RADIO status] The checked attribute is true
PASS [INPUT in RADIO status] The checked attribute is false
PASS [INPUT in RADIO status] The checked attribute is false and the name attribute is empty
PASS [INPUT in FILE status] The required attribute is not set
PASS [INPUT in FILE status] The Files attribute is null
PASS [select]  The required attribute is not set
PASS [select]  Selected the option with value equals to 1
PASS [select]  Selected the option with value equals to empty
PASS [textarea]  The required attribute is not set
PASS [textarea]  The value is not empty
PASS [textarea]  The value is empty
PASS validationMessage should return empty string when willValidate is false and valueMissing is true
Harness: the test ran to completion.
